fix: list finding API
